FMFF/requirement.txt

Django==4.2.5
django-filter==23.3
requests==2.26.0
Pillow==8.3.2
